+++ test/common/definitions.spec.js
@@ -39,4 +39,18 @@ describe('schema definitions', () => {
       }]
     });
   });
+
+  context('address', () => {
+    testValidAndInvalidDefinitions('address', {
+      valid: [{
+        street: '123 a rd',
+        city: 'abc',
+        country: 'USA'
+      }],
+      invalid: [{
+        city: 'foo',
+        country: 'USA'
+      }]
+    });
+  });
 });
